move code to a lua folder
