+++ nvim/init.lua
@@ -15,5 +15,5 @@ if not vim.loop.fs_stat(lazypath) then
 end
 vim.opt.rtp:prepend(lazypath)
 
-require("global-opts")
+require("vim-options")
 require("lazy").setup("plugins")
